requirements.txt
torch
transformers
Pillow
langchain